Bug 422574 - NPE in CoreFilesTest.testMoveFileOverwrite

+++ tests/org.eclipse.orion.server.tests/src/org/eclipse/orion/server/tests/servlets/files/FileSystemTest.java
@@ -16,26 +16,13 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.UnsupportedEncodingException;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URISyntaxException;
+import com.meterware.httpunit.*;
+import java.io.*;
+import java.net.*;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.eclipse.core.filesystem.EFS;
-import org.eclipse.core.filesystem.IFileInfo;
-import org.eclipse.core.filesystem.IFileStore;
-import org.eclipse.core.runtime.CoreException;
-import org.eclipse.core.runtime.IPath;
-import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Path;
-import org.eclipse.core.runtime.Status;
+import org.eclipse.core.filesystem.*;
+import org.eclipse.core.runtime.*;
 import org.eclipse.core.runtime.URIUtil;
 import org.eclipse.orion.internal.server.core.IOUtilities;
 import org.eclipse.orion.internal.server.core.metastore.SimpleMetaStore;
@@ -43,24 +30,12 @@ import org.eclipse.orion.internal.server.core.metastore.SimpleMetaStoreUtil;
 import org.eclipse.orion.internal.server.servlets.ProtocolConstants;
 import org.eclipse.orion.server.core.LogHelper;
 import org.eclipse.orion.server.core.OrionConfiguration;
-import org.eclipse.orion.server.core.metastore.IMetaStore;
-import org.eclipse.orion.server.core.metastore.ProjectInfo;
-import org.eclipse.orion.server.core.metastore.UserInfo;
-import org.eclipse.orion.server.core.metastore.WorkspaceInfo;
+import org.eclipse.orion.server.core.metastore.*;
 import org.eclipse.orion.server.tests.AbstractServerTest;
 import org.eclipse.orion.server.tests.ServerTestsActivator;
 import org.eclipse.orion.server.tests.servlets.internal.DeleteMethodWebRequest;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import org.json.*;
 import org.xml.sax.SAXException;
-
-import com.meterware.httpunit.GetMethodWebRequest;
-import com.meterware.httpunit.PostMethodWebRequest;
-import com.meterware.httpunit.PutMethodWebRequest;
-import com.meterware.httpunit.WebConversation;
-import com.meterware.httpunit.WebRequest;
-import com.meterware.httpunit.WebResponse;
 
 /**
  * Common base class for file system tests.
@@ -196,12 +171,19 @@ public abstract class FileSystemTest extends AbstractServerTest {
 		createFile(makeLocalPathAbsolute(path), fileContent);
 	}
 
+	/**
+	 * Returns an absolute resource URI path within the test project for the given project
+	 * relative path. Returns <code>null</code> if there is no test project configured.
+	 */
 	private String getAbsolutePath(String path) {
 		String userRoot = OrionConfiguration.getUserHome(testUserId).toURI().toString() + "/";
 		String absolutePath;
 		if (OrionConfiguration.getMetaStore() instanceof SimpleMetaStore) {
 			// simple metastore, projects located in user/workspace/project
-			Path basePath = new Path(getTestBaseResourceURILocation());
+			final String baseLocation = getTestBaseResourceURILocation();
+			Path basePath = new Path(baseLocation);
+			if (basePath.segmentCount() < 2)
+				return null;
 			String workspaceId = basePath.segment(0);
 			String workspaceName = SimpleMetaStoreUtil.decodeWorkspaceNameFromWorkspaceId(workspaceId);
 			String projectName = basePath.segment(1);
@@ -230,8 +212,10 @@ public abstract class FileSystemTest extends AbstractServerTest {
 
 	protected void remove(String path) throws CoreException {
 		String absolutePath = getAbsolutePath(path);
-		IFileStore outputFile = EFS.getStore(URI.create(absolutePath));
-		outputFile.delete(EFS.NONE, null);
+		if (absolutePath != null) {
+			IFileStore outputFile = EFS.getStore(URI.create(absolutePath));
+			outputFile.delete(EFS.NONE, null);
+		}
 	}
 
 	private static void transferData(InputStream input, OutputStream output) {
@@ -514,7 +498,10 @@ public abstract class FileSystemTest extends AbstractServerTest {
 	protected WebResponse basicCreateWorkspace(String workspaceName) throws IOException, SAXException {
 		WebRequest request = getCreateWorkspaceRequest(workspaceName);
 		WebResponse response = webConversation.getResponse(request);
-		assertEquals(HttpURLConnection.HTTP_OK, response.getResponseCode());
+		if (response.getResponseCode() != HttpURLConnection.HTTP_OK) {
+			System.err.println(response.getResponseMessage());
+			assertEquals("Unexpected failure creating workspace with name: " + workspaceName, HttpURLConnection.HTTP_OK, response.getResponseCode());
+		}
 		return response;
 
 	}
